I18N-1163 Add Elasticache in front of S3 cache Added Redis pool manager and modified configuration properties
